Fixing the 500 Internal Server Error
The latest version of Werkzeug is using the timezones on the python
datetime objects in its atom feeds. If a datetime object doesn't have
one, it blows up. To fix this, I pinned the version of Werkzeug to the
one that was around when I created this application (0.9). If,
however, you want a better fix for this problem, you should checkout
the better-fix-for-internal-server-error branch. In that branch,
rather than downgrading the version of Werkzeug, I take advantage of
YAML's !!timestamp tag to specify the timezone. This is the prefered
fix for this issue, but I put it on a different branch so the code
works as it was originally detailed in the Tuts+ course.